requirements.txt
langchain==0.2.9
langchainhub==0.1.20
langchain-community==0.2.6
langchain_core==0.2.21
langchain_openai==0.1.17
langchain_text_splitters==0.2.2
openai==1.35.15
SQLAlchemy==2.0.31
streamlit==1.36.0
validators==0.33.0
streamlit-ace
streamlit-toggle
streamlit-toggle-switch
sqlalchemy-vertica-python
